Subindo Atualizações para finalização da api

+++ src/main/java/com/helpdesk/api/service/EquipamentoServiceImpl.java
@@ -5,50 +5,50 @@ import com.helpdesk.api.mapper.EquipamentoMapper;
 import com.helpdesk.api.model.Equipamento;
 import com.helpdesk.api.model.dto.EquipamentoDTO;
 import com.helpdesk.api.repository.EquipamentoRepository;
+
 import com.helpdesk.api.util.MessageConstants;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
 
-import static com.helpdesk.api.mapper.EquipamentoMapper.*;
 
 @RequiredArgsConstructor
 @Service
 public class EquipamentoServiceImpl {
     private final EquipamentoRepository equipamentoRepository;
+    private final EquipamentoMapper equipamentoMapper;
 
     public EquipamentoDTO save(EquipamentoDTO equipamentoDTO) {
-        Equipamento equipamento = INSTANCE.toModel(equipamentoDTO);
+        Equipamento equipamento = equipamentoMapper.toModel(equipamentoDTO);
         Equipamento persistedEquipamento = equipamentoRepository.save(equipamento);
-        return INSTANCE.toDTO(persistedEquipamento);
+        return equipamentoMapper.toDTO(persistedEquipamento);
     }
 
     public List<EquipamentoDTO> findAll() {
         List<Equipamento> equipamentos = equipamentoRepository.findAll();
-        return equipamentos.stream()
-                .map(INSTANCE::toDTO)
-                .collect(Collectors.toList());
+        return equipamentoMapper.toDTOList(equipamentos);
     }
 
-    public Optional<EquipamentoDTO> findById(Long id) {
-        return equipamentoRepository.findById(id)
-                .map(INSTANCE::toDTO);
+    public EquipamentoDTO findById(Long id) {
+        Equipamento equipamento = equipamentoRepository.findById(id)
+                .orElseThrow(() -> new EquipamentoException(MessageConstants.EQUIPAMENTO_COM_O_ID_NAO_EXISTE + id));
+        return equipamentoMapper.toDTO(equipamento);
     }
 
-    public Optional<EquipamentoDTO> update(Long id, EquipamentoDTO equipamentoDTO) {
-        return equipamentoRepository.findById(id)
-                .map(equipamentoAtualizado -> {
-                    Equipamento equipamento = INSTANCE.toModel(equipamentoDTO);
-                    equipamento.setDeviceId(equipamentoAtualizado.getDeviceId());
-                    return INSTANCE.toDTO(equipamentoRepository.save(equipamento));
-                });
+    public EquipamentoDTO update(Long id, EquipamentoDTO equipamentoDTO) {
+        Equipamento equipamento = equipamentoRepository.findById(id)
+                .orElseThrow(() -> new EquipamentoException(MessageConstants.EQUIPAMENTO_COM_O_ID_NAO_EXISTE + id));
+        equipamentoMapper.toUpdate(equipamento, equipamentoDTO);
+        equipamentoRepository.save(equipamento);
+        return equipamentoMapper.toDTO(equipamento);
 
     }
 
     public void delete(Long id) {
+        if (!equipamentoRepository.existsById(id)) {
+            throw new RuntimeException(MessageConstants.EQUIPAMENTO_COM_O_ID_NAO_EXISTE);
+        }
         equipamentoRepository.deleteById(id);
     }
 }
