+++ src/main/java/com/helpdesk/api/service/UsuarioServiceImpl.java
@@ -2,9 +2,8 @@ package com.helpdesk.api.service;
 
 import com.helpdesk.api.exception.UsuarioException;
 import com.helpdesk.api.mapper.UsuarioMapper;
-import com.helpdesk.api.model.Balcao;
 import com.helpdesk.api.model.Usuario;
-import com.helpdesk.api.model.dto.BalcaoDTO;
+
 import com.helpdesk.api.model.dto.UsuarioDTO;
 import com.helpdesk.api.repository.UsuarioRepository;
 import com.helpdesk.api.util.MessageConstants;
@@ -15,39 +14,44 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
 
-import static com.helpdesk.api.mapper.UsuarioMapper.*;
 
 @Service
 @RequiredArgsConstructor
 @Transactional
-public class UsuarioServiceImpl{
+public class UsuarioServiceImpl {
 
     private final UsuarioRepository usuarioRepository;
+    private final UsuarioMapper usuarioMapper;
 
     public UsuarioDTO createUsuario(UsuarioDTO usuarioDTO) {
-        Usuario usuario = INSTANCE.toEntity(usuarioDTO);
+        Usuario usuario = usuarioMapper.toEntity(usuarioDTO);
         Usuario savedUsuario = usuarioRepository.save(usuario);
-        return INSTANCE.toDTO(savedUsuario);
+        return usuarioMapper.toDTO(savedUsuario);
     }
+
     public List<UsuarioDTO> getAllUsuarios() {
         List<Usuario> usuarios = usuarioRepository.findAll();
-        return INSTANCE.usuariosToUsuarioDTO(usuarios);
+        return usuarioMapper.usuariosToUsuarioDTO(usuarios);
     }
+
     public UsuarioDTO getUsuarioById(Long id) {
         Usuario usuario = usuarioRepository.findById(id)
-                .orElseThrow(()-> new UsuarioException(MessageConstants.USUARIO_NAO_ENCONTRADO_COM_ID + id));
-        return INSTANCE.toDTO(usuario);
+                .orElseThrow(() -> new UsuarioException(MessageConstants.USUARIO_NAO_ENCONTRADO_COM_ID + id));
+        return usuarioMapper.toDTO(usuario);
     }
-    public Optional<UsuarioDTO> updateUsuario(Long id, UsuarioDTO updatedUsuarioDTO) {
-        return usuarioRepository.findById(id)
-                .map(existingUsuario -> {
-                    Usuario updatedUsuario = INSTANCE.toEntity(updatedUsuarioDTO);
-                    updatedUsuario.setCustomerId(existingUsuario.getCustomerId());
-                    Usuario savedUsuario = usuarioRepository.save(updatedUsuario);
-                    return INSTANCE.toDTO(savedUsuario);
-                });
+
+    public UsuarioDTO updateUsuario(Long id, UsuarioDTO updatedUsuarioDTO) {
+        Usuario usuario = usuarioRepository.findById(id)
+                .orElseThrow(() -> new UsuarioException(MessageConstants.USUARIO_NAO_ENCONTRADO_COM_ID + id));
+        usuarioMapper.toUpdate(usuario, updatedUsuarioDTO);
+        usuario = usuarioRepository.save(usuario);
+        return usuarioMapper.toDTO(usuario);
     }
+
     public void deleteUsuario(Long id) {
+        if (!usuarioRepository.existsById(id)) {
+            throw new UsuarioException(MessageConstants.USUARIO_NAO_ENCONTRADO_COM_ID + id);
+        }
         usuarioRepository.deleteById(id);
     }
 }
